Issue #956: add AbstractJavaAnnotatableNode class
