+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/AbstractIgnoredAnnotRule.java
@@ -10,7 +10,7 @@ import java.util.Collections;
 import net.sourceforge.pmd.lang.java.ast.Annotatable;
 import net.sourceforge.pmd.properties.StringMultiProperty;
 
-public class AbstractIgnoredAnnotRule extends AbstractJavaRule {
+public abstract class AbstractIgnoredAnnotRule extends AbstractJavaRule {
 
     protected final StringMultiProperty ignoredAnnotationsDescriptor
         = StringMultiProperty.named("ignoredAnnotations")
@@ -22,19 +22,19 @@ public class AbstractIgnoredAnnotRule extends AbstractJavaRule {
         return Collections.emptyList();
     }
 
-    public AbstractIgnoredAnnotRule() {
+    protected AbstractIgnoredAnnotRule() {
         definePropertyDescriptor(ignoredAnnotationsDescriptor);
     }
 
 
     /**
-     * Checks whether any annotation is present on the node.
+     * Checks whether any annotation in ignoredAnnotationsDescriptor is present on the node.
      *
      * @param node
      *            the node to check
      * @return <code>true</code> if the annotation has been found, otherwise <code>false</code>
      */
-    protected boolean isAnyAnnotationPresent(Annotatable node) {
+    protected boolean hasIgnoredAnnotation(Annotatable node) {
         return node.isAnyAnnotationPresent(getProperty(ignoredAnnotationsDescriptor));
     }
 }
