+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/UseUtilityClassRule.java
@@ -18,9 +18,9 @@ import net.sourceforge.pmd.lang.java.ast.ASTMemberValuePair;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTName;
 import net.sourceforge.pmd.lang.java.ast.ASTResultType;
-import net.sourceforge.pmd.lang.java.rule.AbstractLombokAwareRule;
+import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
 
-public class UseUtilityClassRule extends AbstractLombokAwareRule {
+public class UseUtilityClassRule extends AbstractJavaRule {
 
     @Override
     public Object visit(ASTClassOrInterfaceBody decl, Object data) {
@@ -83,30 +83,28 @@ public class UseUtilityClassRule extends AbstractLombokAwareRule {
 
     private boolean isOkUsingLombok(ASTClassOrInterfaceDeclaration parent) {
         // check if there's a lombok no arg private constructor, if so skip the rest of the rules
-        if (hasClassLombokAnnotation()) {
-            ASTAnnotation annotation = parent.getAnnotation("lombok.NoArgsConstructor");
+        ASTAnnotation annotation = parent.getAnnotation("lombok.NoArgsConstructor");
 
-            if (annotation != null) {
+        if (annotation != null) {
 
-                List<ASTMemberValuePair> memberValuePairs = annotation.findDescendantsOfType(ASTMemberValuePair.class);
+            List<ASTMemberValuePair> memberValuePairs = annotation.findDescendantsOfType(ASTMemberValuePair.class);
 
-                for (ASTMemberValuePair memberValuePair : memberValuePairs) {
-                    // to set the access level of a constructor in lombok, you set the access property on the annotation
-                    if ("access".equals(memberValuePair.getImage())) {
-                        List<ASTName> names = memberValuePair.findDescendantsOfType(ASTName.class);
+            for (ASTMemberValuePair memberValuePair : memberValuePairs) {
+                // to set the access level of a constructor in lombok, you set the access property on the annotation
+                if ("access".equals(memberValuePair.getImage())) {
+                    List<ASTName> names = memberValuePair.findDescendantsOfType(ASTName.class);
 
-                        for (ASTName name : names) {
-                            // check to see if the value of the member value pair ends PRIVATE.  This is from the AccessLevel enum in Lombok
-                            if (name.getImage().endsWith("PRIVATE")) {
-                                // if the constructor is found and the accesslevel is private no need to check anything else
-                                return true;
-                            }
+                    for (ASTName name : names) {
+                        // check to see if the value of the member value pair ends PRIVATE.  This is from the AccessLevel enum in Lombok
+                        if (name.getImage().endsWith("PRIVATE")) {
+                            // if the constructor is found and the accesslevel is private no need to check anything else
+                            return true;
                         }
                     }
                 }
             }
         }
-        
+
         return false;
     }
